services/core-api/requirements.txt
# FastAPI and server
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-multipart>=0.0.6,<1.0.0

# Database
sqlalchemy[asyncio]>=2.0.25,<3.0.0
asyncpg>=0.29.0,<1.0.0
alembic>=1.13.0,<2.0.0
greenlet>=3.0.0,<4.0.0

# Redis
redis>=5.0.0,<6.0.0

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.1.0,<5.0.0

# Validation
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
email-validator>=2.1.0,<3.0.0

# HTTP client (for OAuth)
httpx>=0.26.0,<1.0.0

# Rate limiting
slowapi>=0.1.9,<1.0.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
pytest-cov>=4.1.0,<5.0.0
httpx>=0.26.0,<1.0.0
factory-boy>=3.3.0,<4.0.0
faker>=22.0.0,<30.0.0
aiosqlite>=0.19.0,<1.0.0

# Development
ruff>=0.1.0,<1.0.0
